fix(login): strip 55 prefix in phone input
